requirements.txt
addict==2.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
dash==3.0.2
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.57.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.34.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy~=2.1.1
open3d==0.19.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.1
pyparsing==3.2.3
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
PyTurboJPEG==1.7.7
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.0.6
zipp==3.21.0

paramiko~=3.5.1
onnxruntime~=1.21.1